requirements.txt
flask
huggingface-hub
python-dotenv
flask-socketio
eventlet
google-genai
google-cloud-speech
SpeechRecognition
pydub
pyaudio


